Port cargo's libc usage to rustix.
This factors out several `unsafe` blocks, and replaces several instances of
manual error handling with `Result`s.

And it replaces several places that previously manipulated raw file descriptors
to use the I/O safety owning and borrowing types and traits instead.

This also eliminates the need for a special case for musl, allowing cargo to
behave the same way on glibc and musl systems.